pyproject.toml
[project]
name = "langchain-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.2",
    "langchain-core>=1.0.0",
    "langchain-google-vertexai>=3.0.1",
    "python-dotenv>=1.1.1",
]
